remove module_all.ts
