+++ ext/node/lib.rs
@@ -332,7 +332,6 @@ deno_core::extension!(deno_node,
     "internal/util/inspect.mjs",
     "internal/util/types.ts",
     "internal/validators.mjs",
-    "module_all.ts",
     "net.ts",
     "os.ts",
     "path.ts",
